add a book of todos for serving from a folder
also remove zig cache cause its not needed.

The todos are there more to let me think about the architecture i want
to use for this. I am a massive fan of XAMPP and how that worked, so I
want to emulate that but in zig with some of my own constraints (like
always defaulting to an index.html etc)

this is basically nginx and xampp but for total dimwits like me